requirements.txt
torch>=2.0.0
numpy>=1.21.0
optuna>=3.0.0
scipy>=1.7.0
matplotlib>=3.4.0
tqdm>=4.62.0
seaborn>=0.11.0
scikit-learn>=0.24.0
pandas>=1.3.0
umap-learn>=0.5.0
plotly>=5.3.0
kaleido>=0.2.1  # for plotly static image export
flask==3.0.2 